refactor: move user staking data into its own composable which gets consumed by the parent staking provider
