+++ src/components/contextual/pages/pool/StakingIncentivesCard/StakingIncentivesCard.vue
@@ -29,14 +29,16 @@ const stakeAction = ref('');
 const { fNum2 } = useNumbers();
 const { balanceFor } = useTokens();
 const {
-  refetchStakedShares,
-  isStakedSharesIdle,
-  isLoadingStakedShares,
-  isRefetchingStakedShares,
-  isLoadingPoolEligibility,
+  userData: {
+    refetchStakedShares,
+    isStakedSharesIdle,
+    isLoadingStakedShares,
+    isRefetchingStakedShares,
+    stakedSharesForProvidedPool
+  },
   isPoolEligibleForStaking,
-  hideAprInfo,
-  stakedSharesForProvidedPool
+  isLoadingPoolEligibility,
+  hideAprInfo
 } = useStaking();
 
 /**
