+++ src/components/contextual/pages/pools/UnstakedPoolsTable.vue
@@ -24,10 +24,12 @@ const stakePool = ref<FullPool | undefined>();
 
 /** COMPOSABLES */
 const {
-  userGaugeShares,
-  userLiquidityGauges,
-  stakedPools,
-  isLoadingStakingData,
+  userData: {
+    userGaugeShares,
+    userLiquidityGauges,
+    stakedPools,
+    isLoadingUserStakingData
+  },
   setPoolAddress
 } = useStaking();
 
@@ -132,7 +134,7 @@ function handleModalClose() {
     <h5 v-if="!isL2">{{ $t('staking.unstakedPools') }}</h5>
     <PoolsTable
       :key="poolsToRender"
-      :isLoading="isLoadingStakingData || isLoadingUserPools || isUserPoolsIdle"
+      :isLoading="isLoadingUserStakingData || isLoadingUserPools || isUserPoolsIdle"
       :data="poolsToRender"
       :noPoolsLabel="$t('noInvestments')"
       :hiddenColumns="hiddenColumns"
